Added about module and subscriptions
## Added an about module
Added an about module that is required for the bot to run. Tells who created the bot and provides a link to my twitch.

## Added a subscriptions module
Added a subscriptions module that will look for subs and resubs and post a message in chat if one is done. This one hasn't been tested and requires live testing before it can be pushed to stable or dev.

# Notes
Until this branch is confirmed successful, it is **NOT** to be merged into any other branches. Instead, the other branch (dev) should be merged into this one.